refactor: 코드 수정
- 관리자 대시보드 위치 변경
- App.jsx 관리자 방송관리 페이지 이름 수정

FDORA-119
