+++ src/App.jsx
@@ -25,8 +25,7 @@ import SellerHome from './pages/seller/dashboard/SellerHome';
 import LiveSetup from './pages/seller/LiveSetup';
 import Manage from './pages/seller/product/Manage';
 import Register from './pages/seller/product/Register';
-import Adminbroadcast from './pages/adminpage/Adminbroadcast';
-import AdminHome from './pages/adminpage/dashboard/AdminHome';
+import AdminHome from './pages/adminpage/AdminHome';
 import AdminPopup from './pages/adminpage/AdminPopup';
 import AdminPopupRegi from './pages/adminpage/AdminPopupRegi';
 import AdminProduct from './pages/adminpage/AdminProduct';
@@ -37,11 +36,11 @@ import StreamView from './pages/live/StreamView';
 import Event from './pages/event/Event';
 import ChatSupport from './pages/support/ChatSupport';
 import AdminLayout from './layouts/AdminLayout';
-import AdminBroadcast from './pages/adminpage/Adminbroadcast';
+import AdminBroadcast from './pages/adminpage/AdminBroadcast';
 import ProductEdit from './pages/seller/product/ProductEdit';
 import './pages/seller/dashboard/utils/ChartRegistry';
 
-// // import Category from './pages/category/Category';
+// import Category from './pages/category/Category';
 
 function App() {
   return (
